+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- - Split the `property__VariationName` query string into `propertyName` and `propertyValue`.
+
 ## [0.8.0] - 2020-07-10
 
 ### Added